+++ CHANGELOG.md
@@ -285,3 +285,4 @@
   8. Copy menu - fix
   9. Content links removed when removing a content
   10. Generator for Google Analytics script (usage: "rails g mokio:ga_script UA-xxxxxxxx-xx)  - generates Mokio::ExternalScript object and partial view for frontend
+  11. Added frontend helper: build_lang_menu